semester-1/bases-de-datos/petcon/api/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1

mysql-connector-python~=8.0.23